requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3